Quick load rom info and update full romset in the background
Implemented a method where IV-Play loads all of the necessary info to
render the UI immediately on startup/refresh and moved the full romset
info to load in the background. There are still some things to fix
around this but it's a good start. We can now get rid of loading form
completely and provide an accurate percentage for the full update in the
titlebar.