Fix bug #26 - Devices remain in the game list
While I recently added a check to skip adding devices, we never actually removed devices that were already on the gamelist from when we generate the quick gamelist. This commit addresses this both in the database by and the XML parser by deleting anythign that doesn't have a year after updating.